Write the job type "via" in worker info header
so that we can see quickly and via user-viewable logs if the source is HTTP or
AMQP.
